requirements.txt
yfinance==0.2.51
pandas==2.2.3
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
anthropic==0.43.0
requests==2.32.3
click==8.1.8
streamlit==1.41.1
plotly==5.24.1
APScheduler==3.10.4
